Tests for both uap-core versions
