+++ src/test/java/ua_parser/DeviceTest.java
