+++ src/test/java/ua_parser/ParserTestBase.java
@@ -33,15 +33,14 @@ import org.yaml.snakeyaml.Yaml;
  *
  * @author Steve Jiang (@sjiang) <gh at iamsteve com>
  */
-public class ParserTest {
-  final String TEST_RESOURCE_PATH = "/ua_parser/";
+abstract public class ParserTestBase {
   Yaml yaml = new Yaml();
   Parser parser;
 
+  abstract public String resourcePath();
+
   @Before
-  public void initParser() {
-    parser = Parser.newParser();
-  }
+  abstract public void initParser();
 
   @Test
   public void testParseUserAgent() {
@@ -75,22 +74,6 @@ public class ParserTest {
   }
 
   @Test
-  public void testParseAll() {
-    String agentString1 = "Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10.4; fr; rv:1.9.1.5) Gecko/20091102 Firefox/3.5.5,gzip(gfe),gzip(gfe)";
-    String agentString2 = "Mozilla/5.0 (iPhone; CPU iPhone OS 5_1_1 like Mac OS X) AppleWebKit/534.46 (KHTML, like Gecko) Version/5.1 Mobile/9B206 Safari/7534.48.3";
-
-    Client expected1 = new Client(new UserAgent("Firefox", "3", "5", "5"),
-                                  new OS("Mac OS X", "10", "4", null, null),
-                                  new Device("Mac", "Apple", "Mac"));
-    Client expected2 = new Client(new UserAgent("Mobile Safari", "5", "1", null),
-                                  new OS("iOS", "5", "1", "1", null),
-                                  new Device("iPhone", "Apple", "iPhone"));
-
-    assertThat(parser.parse(agentString1), is(expected1));
-    assertThat(parser.parse(agentString2), is(expected2));
-  }
-
-  @Test
   public void testReplacementQuoting() throws Exception {
     String testConfig = "user_agent_parsers:\n"
                       + "  - regex: 'ABC([\\\\0-9]+)'\n"
@@ -115,7 +98,7 @@ public class ParserTest {
   }
 
   void testUserAgentFromYaml(String filename) {
-    InputStream yamlStream = this.getClass().getResourceAsStream(TEST_RESOURCE_PATH + filename);
+    InputStream yamlStream = this.getClass().getResourceAsStream(resourcePath() + filename);
 
     @SuppressWarnings("unchecked")
     Map<String, List<Map<String,String>>> entries = (Map<String, List<Map<String,String>>>)yaml.load(yamlStream);
@@ -131,7 +114,7 @@ public class ParserTest {
   }
 
   void testOSFromYaml(String filename) {
-    InputStream yamlStream = this.getClass().getResourceAsStream(TEST_RESOURCE_PATH + filename);
+    InputStream yamlStream = this.getClass().getResourceAsStream(resourcePath() + filename);
 
     @SuppressWarnings("unchecked")
     Map<String, List<Map<String,String>>> entries = (Map<String, List<Map<String,String>>>)yaml.load(yamlStream);
@@ -147,7 +130,7 @@ public class ParserTest {
   }
 
   void testDeviceFromYaml(String filename) {
-    InputStream yamlStream = this.getClass().getResourceAsStream(TEST_RESOURCE_PATH + filename);
+    InputStream yamlStream = this.getClass().getResourceAsStream(resourcePath() + filename);
 
     @SuppressWarnings("unchecked")
     Map<String, List<Map<String,String>>> entries = (Map<String, List<Map<String,String>>>)yaml.load(yamlStream);
